// src/app/constants.enum.ts
export enum Constants {
    IS_ADMIN = 'isAdmin',
    USER_ID = 'userID',
    TOKEN_KEY = 'auth-token',
    // ROUTE = 'http://localhost:8000',
    ROUTE = 'https://api.paj-gps.com',
    SELECTED_DEVICE = 'selectedDevice',
    DEVICE_NAME = 'deviceName',
    UPDATE_MAP = 'updateMap',
    ROUTE_ICON = 'routeIcon',
    MAP = 'map-style',
    UPDATE_STYLE = 'updateStyle',
    UPDATE_LOGIN = 'updateLogin',
    USER_DATA = 'userData'
}